test event

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Category;
 use App\User;
+use App\CategoryEvent;
 use Auth;
 
 class CategoryController extends Controller
@@ -25,6 +26,14 @@ class CategoryController extends Controller
     }
 
     public function store(Request $request){
+      $rule = [ 'name' => 'required|min:3|max:30'];
+      $message = [
+        'required' => 'Isi bidang ini.',
+        'min' => 'Nama kategori minimal 3 huruf.',
+        'max' => 'Nama kategori maksimal 30 huruf.'
+      ];
+      $this->validate($request, $rule, $message);
+
       $user = User::find(Auth::user()->member_id);
       if ($user->role == "anggota") {
         return response()->json([
@@ -32,16 +41,29 @@ class CategoryController extends Controller
           "status" => false,
         ]);
       }
-      Category::create([
-        "name" => $request->name
-      ]);
-      return response()->json([
-        "message" => "success",
-        "status" => true,
-      ]);
+      $name = ucwords(strtolower($request->name));
+      $categoryName = Category::where('name', $name)->first();
+      if($categoryName){
+        return response()->json([
+          'message' => 'failed',
+          'status' => false,
+          'errors' => 'Nama Kategori '.$name.' sudah ada.'
+        ]);
+      }
+
+      Category::create(["name" => $name]);
+      return response()->json(["message" => "success","status" => true]);
     }
 
     public function update(Request $request, Category $category){
+      $rule = [ 'name' => 'required|min:3|max:30'];
+      $message = [
+        'required' => 'Isi bidang ini.',
+        'min' => 'Nama kategori minimal 3 huruf.',
+        'max' => 'Nama kategori maksimal 30 huruf.'
+      ];
+      $this->validate($request, $rule, $message);
+
       $user = User::find(Auth::user()->member_id);
       if ($user->role == "anggota") {
         return response()->json([
@@ -49,13 +71,23 @@ class CategoryController extends Controller
           "status" => false,
         ]);
       }
-      $category->update([
-        "name" => $request->name
-      ]);
-      return response()->json([
-        "message" => "success",
-        "status" => true,
-      ]);
+
+      $name = ucwords(strtolower($request->name));
+      $categoryName = Category::where('name', $name)->first();
+
+      if($name == $category->name){
+        return response()->json(['message' => 'success','status' => true]);
+      }elseif($categoryName){
+        return response()->json([
+          'message' => 'failed',
+          'status' => false,
+          'errors' => 'Nama Kategori '.$name.' sudah ada.'
+        ]);
+      }
+
+      $category->update(['name' => $name]);
+
+      return response()->json(['message' => 'success','status' => true]);
     }
 
     public function destroy(Category $category){
@@ -66,10 +98,18 @@ class CategoryController extends Controller
           "status" => false,
         ]);
       }
+
+      $categoryEvent = CategoryEvent::where('category_id', $category->id)->first();
+
+      if($categoryEvent){
+        return response()->json([
+          'message' => 'failed',
+          'status' => false,
+          'errors' => 'Kategori '.$categoryEvent->category->name.' sedang digunakan.'
+        ]);
+      }
+
       $category->delete();
-      return response()->json([
-        "message" => "success",
-        "status" => true,
-      ]);
+      return response()->json(['message' => 'success','status' => true]);
     }
 }
